requirements.txt
pydantic
python-dotenv
langchain
langchain-openai
langchain-core
langchain-community
defusedxml
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-neo4j
scikit-learn
neo4j
fastapi
uvicorn
pydantic
python-dotenv
PyPDF2